+++ src/main/java/com/github/leandromoraesrj/sincadbackend/domain/Estabelecimento.java
@@ -1,7 +1,6 @@
 package com.github.leandromoraesrj.sincadbackend.domain;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
@@ -29,7 +28,7 @@ public class Estabelecimento implements Serializable {
 	@SequenceGenerator(name = "se_estabelecimento_generator", sequenceName = "se_estabelecimento", allocationSize = 1)
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "se_estabelecimento_generator")
 	@Column(name = "sq_estabelecimento")
-	private BigDecimal id;
+	private Long id;
 	@Column(name = "nu_cnpj")
 	private String cnpj;
 	@Column(name = "ds_titulo_estabelecimento")
@@ -48,7 +47,7 @@ public class Estabelecimento implements Serializable {
 	public Estabelecimento() {		
 	}
 
-	public Estabelecimento(BigDecimal id, String cnpj, String titulo, UnidadeEmpresarial unidadeEmpresarial) {
+	public Estabelecimento(Long id, String cnpj, String titulo, UnidadeEmpresarial unidadeEmpresarial) {
 		super();
 		this.id = id;
 		this.cnpj = cnpj;
@@ -56,11 +55,11 @@ public class Estabelecimento implements Serializable {
 		this.unidadeEmpresarial = unidadeEmpresarial;
 	}
 
-	public BigDecimal getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(BigDecimal id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
